+++ mass-battle-tracker-reboot-client/src/main/angular/src/app/shared/util/army-by-id.pipe.ts
@@ -1,0 +1,13 @@
+import { Pipe, PipeTransform } from '@angular/core';
+import { Battle } from '../data-model/mass-battle-tracker-reboot-server';
+
+@Pipe({
+  name: 'armyById'
+})
+export class ArmyByIdPipe implements PipeTransform {
+
+  transform(battle: Battle, armyId: number): unknown {
+    return battle.involvedArmies.find(army => army.id === armyId).name;
+  }
+
+}